# packages/spike.land/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Spike Land API
  description: |
    API for Spike Land - AI-Powered App Platform with Image Enhancement

    ## Platform Overview
    Spike Land is a platform where users create, fork, and manage AI-powered applications.
    The flagship application, Pixel, provides AI-powered image enhancement.

    ## Authentication
    All protected endpoints require a valid session cookie from NextAuth.js.
    Alternatively, you can use Bearer token authentication for API requests:
    ```
    Authorization: Bearer {session_token}
    ```

    ## Token Economy
    Users have a platform-wide token balance used for:
    - Image enhancement operations (cost depends on resolution tier)
    - Premium features in apps

    Token costs for Pixel image enhancement:
    - TIER_1K: 2 tokens
    - TIER_2K: 5 tokens
    - TIER_4K: 10 tokens

    Tokens regenerate automatically (1 token per 15 minutes, max 100).

    ## Rate Limiting
    Most endpoints are rate-limited per user. Rate limit information is included in response headers:
    - X-RateLimit-Remaining: Requests remaining in current window
    - X-RateLimit-Reset: Unix timestamp when limit resets

    Example rate limits:
    - Image enhancement: 3 requests per minute
    - Image upload: 10 requests per minute
    - Album operations: 10 requests per minute

    ## Error Handling
    The API uses standard HTTP status codes:
    - 200 OK: Successful request
    - 201 Created: Resource created
...

...

    ## Base URL
...
servers:
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie (automatically set after login)
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for API requests
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Unauthorized"
        code:
          type: string
          description: Machine-readable error code
          example: "UNAUTHORIZED"
        details:
          type: object
          description: Additional error context
          example: { "field": "email", "reason": "invalid format" }
        title:
          type: string
          description: Short error title
          example: "Authentication Failed"
        suggestion:
          type: string
          description: Suggested action to resolve the error
          example: "Please log in again"
    ValidationError:
      type: object
      required:
        - error
        - field
      properties:
        error:
          type: string
          description: Validation error message
          example: "Invalid input"
        field:
          type: string
          description: Field that failed validation
          example: "email"
        code:
          type: string
          enum:
            - REQUIRED_FIELD
            - INVALID_FORMAT
            - FILE_TOO_LARGE
            - UNSUPPORTED_FORMAT
          description: Validation error code
          example: "INVALID_FORMAT"

    # ========================================
    # Enumerations
    # ========================================
    EnhancementTier:
      type: string
      enum:
        - TIER_1K
        - TIER_2K
        - TIER_4K
      description: Image enhancement resolution tier
      example: TIER_2K
...
  responses:
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized"
            code: "UNAUTHORIZED"
            title: "Authentication Failed"
            suggestion: "Please log in again"
    Forbidden:
      description: User lacks permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Insufficient permissions"
            code: "FORBIDDEN"
            title: "Access Denied"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            title: "Not Found"
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Error"
              - $ref: "#/components/schemas/ValidationError"
          example:
            error: "Invalid input"
            code: "INVALID_INPUT"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
            title: "Too Many Requests"
            suggestion: "Please wait before making another request"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error"
            code: "INTERNAL_SERVER_ERROR"
            title: "Server Error"
            suggestion: "Please try again later"

  # ========================================
  # Request Parameters
  # ========================================
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      description: Resource ID (CUID)
      schema:
        type: string
      example: "cuid123456789"
...
paths:
...
  /api/albums/{id}/enhance:
    post:
      summary: Batch enhance album images
      description: |
        Batch enhance all or specific images in an album with a specified enhancement tier.
        This is a resource-intensive operation that consumes tokens from user balance.
        Maximum 20 images per batch. Skips already enhanced images by default.
      operationId: enhanceAlbumImages
      tags:
        - Albums
        - Enhancement
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        description: Batch enhancement request
        content:
          application/json:
            schema:
              type: object
              required:
                - tier
              properties:
                tier:
                  $ref: "#/components/schemas/EnhancementTier"
                skipAlreadyEnhanced:
                  type: boolean
                  default: true
                  description: Skip images that already have enhancements at this tier
            example:
              tier: "TIER_2K"
              skipAlreadyEnhanced: true
      responses:
        "200":
          description: Batch enhancement queued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  totalImages:
                    type: integer
                  skipped:
                    type: integer
                  queued:
                    type: integer
                  totalCost:
                    type: integer
                  newBalance:
                    type: integer
                  jobs:
                    type: array
                    items:
                      type: object
                      properties:
                        imageId:
                          type: string
                        jobId:
                          type: string
              example:
                success: true
                totalImages: 5
                skipped: 2
                queued: 3
                totalCost: 15
                newBalance: 85
                jobs:
                  - imageId: "img_001"
                    jobId: "album-clv8k9r9w000108jp5x4y5k8z-1702237500000"
                  - imageId: "img_003"
                    jobId: "album-clv8k9r9w000108jp5x4y5k8z-1702237500000"
                  - imageId: "img_005"
                    jobId: "album-clv8k9r9w000108jp5x4y5k8z-1702237500000"
        "400":
          description: Invalid tier or batch exceeds size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Maximum 20 images allowed per batch enhancement. This album has 25 images to enhance."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          description: Insufficient tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  required:
                    type: integer
                  toEnhance:
                    type: integer
              example:
                error: "Insufficient tokens"
                required: 100
                toEnhance: 10
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Album not found or no images in album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "No images found in album"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/images/enhance:
    post:
      summary: Enhance a single image
      tags: [Enhancement]
      operationId: enhanceImage
      description: |
        Start an AI enhancement job for a single image. The enhancement is processed
        asynchronously and you can poll the job status using the returned jobId.
        Consumes tokens based on selected tier. Operation fails if insufficient tokens.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageId, tier]
              properties:
                imageId:
                  type: string
                  description: ID of the image to enhance
                tier:
                  $ref: "#/components/schemas/EnhancementTier"
      responses:
        "200":
          description: Enhancement job started successfully
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [success, jobId, tokenCost, newBalance]
                properties:
                  success:
                    type: boolean
                  jobId:
                    type: string
                    description: Enhancement job ID for tracking progress
                  tokenCost:
                    type: integer
                    description: Tokens consumed for this enhancement
                  newBalance:
                    type: integer
                    description: User's remaining token balance after consumption
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          description: Insufficient tokens for this operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
...